app/requirements.txt
Flask==2.3.3
google-cloud-vision==3.5.0
google-cloud-firestore==2.13.0
